email_llm_api/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
openai
imaplib2
email-validator
python-dotenv
pandas
numpy
faiss-cpu
sentence-transformers
tiktoken
beautifulsoup4
html2text
python-multipart
jinja2
aiofiles
pydantic
httpx
asyncio
typing-extensions